requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
httpx==0.25.1
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.5
sendgrid==6.10.0
twilio==8.10.0
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
gunicorn==21.2.0
# ChatterBot==1.0.8
# spacy==3.7.2
